Check initial values for mixglm and generate new ones if necessary

This function checks initial values for mixglm and if they are illegitimate, it tries to generate new ones.

Usage

checkInit(
  stateValModels,
  statePrecModels,
  inputData,
  setInit,
  errorModel,
  linkFunction,
  Nstates,
  genNew = FALSE,
  genFn = function(x) rnorm(1, 0, 2),
  checkForSimulation = FALSE
)

Arguments

stateValModels

A formula describing the regression relationship between the mean of the response variable and the covariates for all mixture components, or a list of formulas with each element giving the regression relationship between the mean of the response variable and the covariates for each mixture component (ordered according to intercept of the mixture component on the y-axis). The response variable must be given on the left-hand side of the formula.

statePrecModels

A formula describing the regression relationship between the variances in the mixture components and the covariates, or a list of formulas with each mixture component giving the regression relationship between the variance in the mixture component and the covariates for each mixture component (ordered according to intercept of the mixture component on the y-axis).

inputData

A data frame containing the covariate information.

setInit

List of initial values. It is assumed that for each parameter, there is a vector of initial values with length corresponding to Nstates

errorModel

A string specifying the error distribution to be used in the mixglm. This can be: "gaussian", "gamma", "beta" or "negbinomial".

linkFunction

A string specifying the link function. This can be: "identity", "log", "logit", "probit", or "cloglog".

Nstates

An integer specifying number of components in the mixture.

genNew

A logical value specifying if new initial values should be generated even if setInit contains legitimate values.

genFn

A function used to generate new initial values. It has one argument which is the component rank and it should return one value. Absolute value of provided function is used for initial value of intercept of mean for components apart from the first one. Note that for intercept of precision of beta distribution, genFn(x) + 6 is used.

checkForSimulation

A logical value specifying if only a check of initial values should be done. Mainly for internal use.

Value

A list of initial values

Author

Adam Klimes